See extended description
Previous commit has bad message because I forgot to commit those changes
and needed to call my brother to do the commit for me.

What happened in the last commit was integration of new sliders into
FilmFinder.
Fixed a few bugs in the ActiveArea slider where the red triangle would
return to the center of the slider upon MouseLeave and MouseClick.

I believe that's it.