Fixed bug with active area
Bug was that RangeOfValues would not update during mousewheel events.

This can be seen by navigating by mouse to a low density index in
Directors and then using the mouse wheel to get the the next index. You
will see that the red slider does not move along the main slider, this
is because the RangeOfValues does not update properly.

The fix was to change
updateRangeOfValues(tempValue)
to
base.updateOffset(tempValue)

in the OnMouseWheel events for both DD and IDActiveArea slider