Movie filtering defferal. IDSlider changes.
Movie filtering defferal worked. UI feels snappier. Had to make some
other subtle changes. MovieHandler.cs now raises an event when the
movies are filtered and FilmFinderGUI.cs listens for that event.
Previously "drawActiveSet" would be called in FilmFinderGUI.cs everytime
a filter was updated but now it's only called when the new event from
MovieHandler.cs is raised.

Minor change to IDSlider - if we have lots of As or Zs the slider won't
go out of bounds anymore but will shrink accordingly.