Added training mode
This new mode is handled behind the scenes by saving and setting an
application setting. The change is made upon experiment completion.

If you're in the wrong mode you need to "complete" the experiment to get
to the next mode.

The title of the form indicates the mode you're in.